Implemented Theme Changer - Added Settings Activity
