+++ app/src/main/java/com/viiishoppinglistapp/doit/ChangeThemeActivity.java
@@ -1,14 +1,51 @@
 package com.viiishoppinglistapp.doit;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.CompoundButton;
+import android.widget.Switch;
 
 public class ChangeThemeActivity extends AppCompatActivity {
+
+    Switch aSwitch;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_change_theme);
+
+        setupSwitch();
     }
+
+    private void setupSwitch() {
+        aSwitch = findViewById(R.id.switchThemeMode);
+        if(AppCompatDelegate.getDefaultNightMode()==AppCompatDelegate.MODE_NIGHT_YES){
+            aSwitch.setChecked(true);
+        }
+        aSwitch.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
+            @Override
+            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
+                if(isChecked){
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
+                }
+                else {
+                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
+                }
+            }
+        });
+    }
+
+    public void goToHome(View view){
+        //goto Home page
+        Bundle bundle = new Bundle();
+        bundle.putString("list_name", "Default");
+        Intent I = new Intent(this, TabbedHomeActivity.class);
+        I.putExtras(bundle);
+        this.startActivity(I);
+    }
+
 }
