+++ app/src/main/res/layout/activity_change_theme.xml
@@ -73,7 +73,7 @@
                 android:layout_height="wrap_content"
                 android:layout_marginHorizontal="40dp"
                 android:layout_marginTop="50dp"
-                android:textAlignment="center" />
+                android:text="Enable Dark Mode"/>
 
         </LinearLayout>
